+++ src/Exception/ItemTooLargeException.php
@@ -7,7 +7,7 @@
  */
 declare(strict_types=1);
 
-namespace DVDoug\BoxPacker;
+namespace DVDoug\BoxPacker\Exception;
 
 /**
  * Class ItemTooLargeException